reworked code
