+++ src/components/index.ts
@@ -1,2 +1,3 @@
 export * from './LocalePickerToolbarPlugin';
 export * from './InlineTranslatedText';
+export * from './ModalBuilder';
