+++ src/index.ts
@@ -1,3 +1,4 @@
 export * from './components';
 export * from './localization-api';
 export * from './hooks';
+export * from './plugins';
